--- SimplePolicy/PolicyType.cs
@@ -1,10 +1,0 @@
-﻿namespace Neo.Plugins
-{
-    internal enum PolicyType : byte
-    {
-        AllowAll,
-        DenyAll,
-        AllowList,
-        DenyList
-    }
-}
