+++ SimplePolicy/Settings.cs
@@ -1,8 +1,9 @@
-﻿using System;
+using Microsoft.Extensions.Configuration;
+using Neo.Wallets;
+using System;
 using System.Collections.Generic;
 using System.Linq;
-using Microsoft.Extensions.Configuration;
-using Neo.Network.P2P.Payloads;
+using System.Reflection;
 
 namespace Neo.Plugins
 {
@@ -15,16 +16,21 @@ namespace Neo.Plugins
         public BlockedAccounts BlockedAccounts { get; }
         public HashSet<TransactionType> HighPriorityTx { get; set; }
 
-        public static Settings Default { get; private set; }
+        public static Settings Default { get; }
 
-        private Settings(IConfigurationSection section)
+        static Settings()
         {
-            MaxTransactionsPerBlock = GetValueOrDefault(section.GetSection("MaxTransactionsPerBlock"), 500, p => int.Parse(p));
-            MaxFreeTransactionsPerBlock = GetValueOrDefault(section.GetSection("MaxFreeTransactionsPerBlock"), 20, p => int.Parse(p));
-            MaxFreeTransactionSize = GetValueOrDefault(section.GetSection("MaxFreeTransactionSize"), 1024, p => int.Parse(p));
-            FeePerExtraByte = GetValueOrDefault(section.GetSection("FeePerExtraByte"), Fixed8.FromDecimal(0.00001M), p => Fixed8.Parse(p));
-            BlockedAccounts = new BlockedAccounts(section.GetSection("BlockedAccounts"));
-            HighPriorityTx = new HashSet<TransactionType>(section.GetSection("HighPriorityTx").GetChildren()
+            Default = new Settings(Assembly.GetExecutingAssembly().GetConfiguration());
+        }
+
+        public Settings(IConfigurationSection section)
+        {
+            this.MaxTransactionsPerBlock = GetValueOrDefault(section.GetSection("MaxTransactionsPerBlock"), 500, p => int.Parse(p));
+            this.MaxFreeTransactionsPerBlock = GetValueOrDefault(section.GetSection("MaxFreeTransactionsPerBlock"), 20, p => int.Parse(p));
+            this.MaxFreeTransactionSize = GetValueOrDefault(section.GetSection("MaxFreeTransactionSize"), 1024, p => int.Parse(p));
+            this.FeePerExtraByte = GetValueOrDefault(section.GetSection("FeePerExtraByte"), Fixed8.FromDecimal(0.00001M), p => Fixed8.Parse(p));
+            this.BlockedAccounts = new BlockedAccounts(section.GetSection("BlockedAccounts"));
+            this.HighPriorityTx = new HashSet<TransactionType>(section.GetSection("HighPriorityTx").GetChildren()
                 .Select(p => (TransactionType)Enum.Parse(typeof(TransactionType), p.Value)));
         }
 
@@ -37,6 +43,26 @@ namespace Neo.Plugins
         public static void Load(IConfigurationSection section)
         {
             Default = new Settings(section);
+        }
+    }
+
+    internal enum PolicyType : byte
+    {
+        AllowAll,
+        DenyAll,
+        AllowList,
+        DenyList
+    }
+
+    internal class BlockedAccounts
+    {
+        public PolicyType Type { get; }
+        public HashSet<UInt160> List { get; }
+
+        public BlockedAccounts(IConfigurationSection section)
+        {
+            this.Type = section.GetSection("Type").GetValueOrDefault(PolicyType.AllowAll, p => (PolicyType)Enum.Parse(typeof(PolicyType), p, true));
+            this.List = new HashSet<UInt160>(section.GetSection("List").GetChildren().Select(p => p.Value.ToScriptHash()));
         }
     }
 }
